фикс ошибок в загрузке прайс-листов, дозагрузка изображений на превью шаблона

+++ EtkBlazorApp.BL/Managers/PriceListManager.cs
@@ -4,7 +4,7 @@ using EtkBlazorApp.Core.Data;
 using EtkBlazorApp.DataAccess;
 using EtkBlazorApp.DataAccess.Entity.PriceList;
 using EtkBlazorApp.DataAccess.Repositories.PriceList;
-using Microsoft.Extensions.Logging;
+using NLog;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -18,6 +18,8 @@ namespace EtkBlazorApp.BL.Managers
     //TODO: разбить/упростить класс
     public class PriceListManager
     {
+        private static readonly Logger nlog = LogManager.GetCurrentClassLogger();
+
         public event Action OnPriceListLoaded;
         public decimal NDS { get; private set; }
         public List<PriceLine> PriceLines { get; }
@@ -26,19 +28,16 @@ namespace EtkBlazorApp.BL.Managers
         private readonly IPriceLineLoadCorrelator correlator;
         private readonly IPriceListTemplateStorage templateStorage;
         private readonly ISettingStorageReader settings;
-        private readonly ILogger<PriceListManager> logger;
 
         public PriceListManager(IPriceLineLoadCorrelator correlator,
             IPriceListTemplateStorage templateStorage,
-            ISettingStorageReader settings,
-            ILogger<PriceListManager> logger)
+            ISettingStorageReader settings)
         {
             PriceLines = new List<PriceLine>();
             LoadedFiles = new List<LoadedPriceListTemplateData>();
             this.correlator = correlator;
             this.templateStorage = templateStorage;
             this.settings = settings;
-            this.logger = logger;
         }
 
         /// <summary>
@@ -50,8 +49,12 @@ namespace EtkBlazorApp.BL.Managers
         /// <returns></returns>
         public async Task UploadTemplate(Type templateType, Stream stream, string fileName)
         {
+            nlog.Info("Начало загрузки прайс-листа вида {templateName} из файла {fileName}", templateType.Name, fileName);
+
             var data = await ReadTemplateLines(templateType, stream, fileName, addFileData: true);
             AddNewPriceLines(data);
+
+            nlog.Info("Загружены данные для {count} товаров", data?.Count ?? 0);
         }
 
         /// <summary>
@@ -74,15 +77,22 @@ namespace EtkBlazorApp.BL.Managers
 
             try
             {
+                nlog.Info("Начало загрузки шаблона типа {templateTypeName} из файла {fileName} по временному пути {tempPath}"
+                    , templateType.Name, fileName, tempFilePath);
                 list = await ReadPriceLinesAndApplyModifiers(templateType, stream, fileName, addFileData, tempFilePath);
             }
-            catch
+            catch (Exception ex)
             {
+                nlog.Warn("Ошибка загрузки шаблона {templateTypeName}. Message: {msg}. StackTrace: {stackTrace}", templateType.Name, ex.Message, ex.StackTrace);
                 throw;
             }
             finally
             {
-                File.Delete(tempFilePath);
+                if (File.Exists(tempFilePath))
+                {
+                    nlog.Trace("Временны файл {tempPath} удален", tempFilePath);
+                    File.Delete(tempFilePath);
+                }
             }
 
             return list;
@@ -109,6 +119,7 @@ namespace EtkBlazorApp.BL.Managers
 
             if (templateInfo == null)
             {
+                nlog.Warn("templateInfo пуст для типа {typeName}", templateType.Name);
                 throw new ArgumentNullException($"'{nameof(templateInfo)}' не может быть null");
             }
 
@@ -125,7 +136,9 @@ namespace EtkBlazorApp.BL.Managers
                 }
             }
 
+            nlog.Trace("Начало загрузки данных из шаблона прайс-листа типа {typeName}", templateType.Name);
             var list = await templateInstance.ReadPriceLines(CancellationToken.None);
+            nlog.Trace("Загружены данные для {count} товаров из шаблона прайс-листа типа {typeName}", list.Count, templateType.Name);
 
             await ApplyModifiers(list, templateInfo);
 
@@ -146,7 +159,7 @@ namespace EtkBlazorApp.BL.Managers
         /// <returns></returns>
         private async Task ApplyModifiers(List<PriceLine> list, PriceListTemplateEntity templateInfo)
         {
-            SkipUnnecessaryManufacturers(list, templateInfo);
+            MapManufacturerNamesAndSkipUnnecessaryManufacturers(list, templateInfo);
 
             FillLinkedStock(list, templateInfo.stock_partner_id);
 
@@ -161,30 +174,78 @@ namespace EtkBlazorApp.BL.Managers
         /// <param name="templateInfo"></param>
         /// <param name="list"></param>
         /// <exception cref="NotImplementedException"></exception>
-        private void SkipUnnecessaryManufacturers(List<PriceLine> list, PriceListTemplateEntity templateInfo)
+        private void MapManufacturerNamesAndSkipUnnecessaryManufacturers(List<PriceLine> list, PriceListTemplateEntity templateInfo)
         {
+            if (list is null || templateInfo is null)
+            {
+                return;
+            }
+
+            if ((templateInfo?.manufacturer_name_map?.Count ?? 0) > 0)
+            {
+                Dictionary<string, string> brandNameMap = templateInfo.manufacturer_name_map
+                    .Where(i => !string.IsNullOrWhiteSpace(i.text) && !string.IsNullOrWhiteSpace(i.name))
+                    .ToDictionary(i => i.text, i => i.name);
+
+                if ((list != null && list.Any()) && (brandNameMap != null && brandNameMap.Any()))
+                {
+                    foreach (var line in list)
+                    {
+                        if (!string.IsNullOrWhiteSpace(line.Manufacturer) && brandNameMap.ContainsKey(line.Manufacturer))
+                        {
+                            line.Manufacturer = brandNameMap[line.Manufacturer];
+                        }
+                    }
+                }
+            }
+
             if ((templateInfo?.manufacturer_skip_list?.Count ?? 0) == 0)
             {
                 return;
             }
 
+            int linesAtStart = list.Count;
+            int whiteListDeleted = 0;
+            int blackListDeleted = 0;
+
             //Применяем черный лист
-            var blackList = templateInfo.manufacturer_skip_list.Where(i => i.list_type == "black_list").ToArray();
-            if (blackList.Length > 0)
+            var blackList = templateInfo.manufacturer_skip_list
+                .Where(i => i.list_type == "black_list" && !string.IsNullOrWhiteSpace(i.name))
+                .Select(i => i.name)
+                .ToHashSet();
+            if (blackList.Count > 0)
             {
-                var forbiddenManufacturerNames = blackList.Select(m => m.name).ToArray();
-                list.RemoveAll(priceLine => forbiddenManufacturerNames.Contains(priceLine.Manufacturer, StringComparer.OrdinalIgnoreCase));
+                list.RemoveAll(priceLine => blackList.Contains(priceLine.Manufacturer, StringComparer.OrdinalIgnoreCase));
+                foreach (var line in list.ToArray())
+                {
+                    if (!string.IsNullOrWhiteSpace(line.Manufacturer) && blackList.Contains(line.Manufacturer, StringComparer.OrdinalIgnoreCase))
+                    {
+                        list.Remove(line);
+                    }
+                }
+                blackListDeleted = list.Count - linesAtStart;
             }
 
             //Применяем белый лист
-            var whiteList = templateInfo.manufacturer_skip_list.Where(i => i.list_type == "white_list").ToArray();
-            if (whiteList.Length > 0)
+            var whiteList = templateInfo.manufacturer_skip_list
+                .Where(i => i.list_type == "white_list" && !string.IsNullOrWhiteSpace(i.name))
+                .Select(i => i.name)
+                .ToHashSet();
+            if (whiteList.Count > 0)
             {
-                var allowedManufacturerNames = whiteList.Select(m => m.name).ToArray();
-                list.RemoveAll(priceLine => !allowedManufacturerNames.Contains(priceLine.Name, StringComparer.OrdinalIgnoreCase));
-
+                foreach (var line in list.ToArray())
+                {
+                    if (!string.IsNullOrWhiteSpace(line.Manufacturer) && !whiteList.Contains(line.Manufacturer, StringComparer.OrdinalIgnoreCase))
+                    {
+                        list.Remove(line);
+                    }
+                }
+                whiteListDeleted = linesAtStart - list.Count - blackListDeleted;
             }
 
+            int deletedCount = linesAtStart - list.Count;
+            nlog.Trace("Удаление данных для ненужных брендов. Было {countBefore} стало {countAfter} записей. Удалено всего {deletedCount} (из них белый список: {whiteList}, черный список: {blackList})"
+                , linesAtStart, list.Count, deletedCount, whiteListDeleted, blackListDeleted);
         }
 
         /// <summary>
@@ -220,7 +281,8 @@ namespace EtkBlazorApp.BL.Managers
 
             var modelMap = templateInfo.model_map.ToDictionary(i => i.old_text, i => i.new_text);
 
-            logger.LogInformation($"Запуск преобразования моделей/артикулов для прайс-листа '{templateInfo?.title ?? "<Пусто>"}'");
+            nlog.Trace("Запуск преобразования моделей/артикулов для прайс-листа '{templateTitle}'"
+                , templateInfo?.title);
 
             var sw = new Stopwatch();
 
@@ -242,7 +304,8 @@ namespace EtkBlazorApp.BL.Managers
                 }
             }
 
-            logger.LogInformation($"Конец преобразования моделей/артикулов для прайс-листа '{templateInfo?.title ?? "<Пусто>"}'. Длительность выполнения: {sw.Elapsed.TotalMilliseconds:N} ms");
+            nlog.Trace("Конец преобразования моделей/артикулов для прайс-листа '{templateTitle}'. Длительность выполнения: {elapsedMs} ms",
+                templateInfo?.title, (int)sw.Elapsed.TotalMilliseconds);
         }
 
         /// <summary>
@@ -309,7 +372,7 @@ namespace EtkBlazorApp.BL.Managers
 
             decimal discountValue = emptyDiscountGeneralDiscount ? 0m : templateInfo.discount;
             // Если бренд есть в словаре, то берем скидку для него в первую очередь (заменяя стандартную от прайс-листа)
-            if (!emptySellDiscountMap && sellDiscounts.ContainsKey(line.Manufacturer))
+            if (!emptySellDiscountMap && !string.IsNullOrWhiteSpace(line.Manufacturer) && sellDiscounts.ContainsKey(line.Manufacturer))
             {
                 discountValue = sellDiscounts[line.Manufacturer];
             }
